requirements-desktop.txt
streamlit>=1.31
Pillow>=10.0
reportlab>=4.0
toml>=0.10
